Address SonarLint changes

+++ 02_Inventory/src/main/java/inventory/controller/MainScreenController.java
@@ -23,12 +23,7 @@ import java.util.ResourceBundle;
 
 
 public class MainScreenController implements Initializable,Controller {
-    
-     // Declare fields
-    private Stage stage;
-    private Parent scene;
-    private static Part modifyPart;
-    private static Product modifyProduct;
+
     private static int modifyPartIndex;
     private static int modifyProductIndex;
     
@@ -80,8 +75,6 @@ public class MainScreenController implements Initializable,Controller {
     @FXML
     private TextField productsSearchTxt;
 
-    public MainScreenController(){}
-
     public void setService(InventoryService service){
         this.service=service;
         partsTableView.setItems(service.getAllParts());
@@ -114,10 +107,11 @@ public class MainScreenController implements Initializable,Controller {
      * @throws IOException
      */
     private void displayScene(ActionEvent event, String source) throws IOException {
-        stage = (Stage)((Button)event.getSource()).getScene().getWindow();
+        // Declare fields
+        Stage stage = (Stage) ((Button) event.getSource()).getScene().getWindow();
         FXMLLoader loader= new FXMLLoader(getClass().getResource(source));
         //scene = FXMLLoader.load(getClass().getResource(source));
-        scene = loader.load();
+        Parent scene = loader.load();
         Controller ctrl=loader.getController();
         ctrl.setService(service);
         stage.setScene(new Scene(scene));
@@ -139,11 +133,13 @@ public class MainScreenController implements Initializable,Controller {
         alert.setContentText("Are you sure you want to delete part " + part.getName() + " from parts?");
         Optional<ButtonType> result = alert.showAndWait();
 
-        if (result.get() == ButtonType.OK) {
-            System.out.println("Part deleted.");
-            service.deletePart(part);
-        } else {
-            System.out.println("Canceled part deletion.");
+        if (result.isPresent()) {
+            if (result.get() == ButtonType.OK) {
+                System.out.println("Part deleted.");
+                service.deletePart(part);
+            } else {
+                System.out.println("Canceled part deletion.");
+            }
         }
     }
 
@@ -198,7 +194,7 @@ public class MainScreenController implements Initializable,Controller {
      */
     @FXML
     void handleModifyPart(ActionEvent event) throws IOException {
-        modifyPart = partsTableView.getSelectionModel().getSelectedItem();
+        Part modifyPart = partsTableView.getSelectionModel().getSelectedItem();
         modifyPartIndex = service.getAllParts().indexOf(modifyPart);
         
         displayScene(event, "/fxml/ModifyPart.fxml");
@@ -211,7 +207,7 @@ public class MainScreenController implements Initializable,Controller {
      */
     @FXML
     void handleModifyProduct(ActionEvent event) throws IOException {
-        modifyProduct = productsTableView.getSelectionModel().getSelectedItem();
+        Product modifyProduct = productsTableView.getSelectionModel().getSelectedItem();
         modifyProductIndex = service.getAllProducts().indexOf(modifyProduct);
         
         displayScene(event, "/fxml/ModifyProduct.fxml");
